Split `Groups.kt` to multiple files
