+++ markdown/markdown-backend/src/main/kotlin/party/markdown/backend/groups/Outbox+Operators.kt
@@ -1,0 +1,6 @@
+package party.markdown.backend.groups
+
+/** Maps the given [Outbox] using the provided function. */
+fun <A, B> Outbox<A>.map(
+    f: suspend (B) -> A,
+) = Outbox<B> { element -> this@map.sendCatching(f(element)) }
